requirements.txt
python-dotenv

telethon==1.39

db-sqlite3